v0.2.0: WALV built with Vuejs
